Tidy up the repository for scope of library
