+++ include/boost/capy.hpp
@@ -37,21 +37,16 @@
 #include <boost/capy/detail/config.hpp>
 #include <boost/capy/ex/any_coro.hpp>
 #include <boost/capy/datastore.hpp>
-#include <boost/capy/core/embed.hpp>
 #include <boost/capy/error.hpp>
 #include <boost/capy/ex/execution_context.hpp>
 #include <boost/capy/ex/executor_work_guard.hpp>
 #include <boost/capy/file.hpp>
 #include <boost/capy/file_mode.hpp>
 #include <boost/capy/ex/frame_allocator.hpp>
-#include <boost/capy/core/intrusive_list.hpp>
-#include <boost/capy/core/intrusive_queue.hpp>
 #include <boost/capy/ex/make_affine.hpp>
-#include <boost/capy/core/neunique_ptr.hpp>
 #include <boost/capy/core/polystore.hpp>
 #include <boost/capy/core/polystore_fwd.hpp>
 #include <boost/capy/ex/run_on.hpp>
-#include <boost/capy/core/small_unique_ptr.hpp>
 #include <boost/capy/ex/strand.hpp>
 #include <boost/capy/task.hpp>
 #include <boost/capy/core/thread_local_ptr.hpp>
