+++ include/boost/capy/detail/intrusive.hpp
@@ -7,34 +7,20 @@
 // Official repository: https://github.com/cppalliance/capy
 //
 
-#ifndef BOOST_CAPY_INTRUSIVE_LIST_HPP
-#define BOOST_CAPY_INTRUSIVE_LIST_HPP
-
-#include <boost/capy/detail/config.hpp>
+#ifndef BOOST_CAPY_DETAIL_INTRUSIVE_HPP
+#define BOOST_CAPY_DETAIL_INTRUSIVE_HPP
 
 namespace boost {
 namespace capy {
+namespace detail {
+
+//------------------------------------------------
 
 /** An intrusive doubly linked list.
 
     This container provides O(1) push and pop operations for
     elements that derive from @ref node. Elements are not
     copied or moved; they are linked directly into the list.
-
-    @par Usage
-    @code
-    struct my_item : intrusive_list<my_item>::node
-    {
-        // user data
-    };
-
-    using item_list = intrusive_list<my_item>;
-
-    my_item item;
-    item_list q;
-    q.push_back(&item);
-    my_item* p = q.pop_front();  // p == &item
-    @endcode
 
     @tparam T The element type. Must derive from `intrusive_list<T>::node`.
 */
@@ -62,23 +48,8 @@ private:
     T* tail_ = nullptr;
 
 public:
-    /** Default constructor.
-
-        Creates an empty list.
-
-        @post `empty() == true`
-    */
     intrusive_list() = default;
 
-    /** Move constructor.
-
-        Takes ownership of all elements from `other`,
-        leaving `other` empty.
-
-        @param other The list to move from.
-
-        @post `other.empty() == true`
-    */
     intrusive_list(intrusive_list&& other) noexcept
         : head_(other.head_)
         , tail_(other.tail_)
@@ -91,22 +62,12 @@ public:
     intrusive_list& operator=(intrusive_list const&) = delete;
     intrusive_list& operator=(intrusive_list&&) = delete;
 
-    /** Return true if the list is empty.
-
-        @return `true` if the list contains no elements.
-    */
     bool
     empty() const noexcept
     {
         return head_ == nullptr;
     }
 
-    /** Add an element to the back of the list.
-
-        @param w Pointer to the element to add.
-
-        @pre `w` is not null and not already in a list.
-    */
     void
     push_back(T* w) noexcept
     {
@@ -119,15 +80,6 @@ public:
         tail_ = w;
     }
 
-    /** Splice all elements from another list to the back.
-
-        All elements from `other` are moved to the back of this
-        list. After this call, `other` is empty.
-
-        @param other The list to splice from.
-
-        @post `other.empty() == true`
-    */
     void
     splice_back(intrusive_list& other) noexcept
     {
@@ -148,11 +100,6 @@ public:
         other.tail_ = nullptr;
     }
 
-    /** Remove and return the front element.
-
-        @return Pointer to the front element, or `nullptr`
-            if the list is empty.
-    */
     T*
     pop_front() noexcept
     {
@@ -167,15 +114,6 @@ public:
         return w;
     }
 
-    /** Remove a specific element from the list.
-
-        Unlinks the given element from its current position
-        in the list. The element must be a member of this list.
-
-        @param w Pointer to the element to remove.
-
-        @pre `w` is not null and is currently in this list.
-    */
     void
     remove(T* w) noexcept
     {
@@ -190,6 +128,102 @@ public:
     }
 };
 
+//------------------------------------------------
+
+/** An intrusive singly linked FIFO queue.
+
+    This container provides O(1) push and pop operations for
+    elements that derive from @ref node. Elements are not
+    copied or moved; they are linked directly into the queue.
+
+    Unlike @ref intrusive_list, this uses only a single `next_`
+    pointer per node, saving memory at the cost of not supporting
+    O(1) removal of arbitrary elements.
+
+    @tparam T The element type. Must derive from `intrusive_queue<T>::node`.
+*/
+template<class T>
+class intrusive_queue
+{
+public:
+    /** Base class for queue elements.
+
+        Derive from this class to make a type usable with
+        @ref intrusive_queue. The `next_` pointer is private
+        and accessible only to the queue.
+    */
+    class node
+    {
+        friend class intrusive_queue;
+
+    private:
+        T* next_;
+    };
+
+private:
+    T* head_ = nullptr;
+    T* tail_ = nullptr;
+
+public:
+    intrusive_queue() = default;
+
+    intrusive_queue(intrusive_queue&& other) noexcept
+        : head_(other.head_)
+        , tail_(other.tail_)
+    {
+        other.head_ = nullptr;
+        other.tail_ = nullptr;
+    }
+
+    intrusive_queue(intrusive_queue const&) = delete;
+    intrusive_queue& operator=(intrusive_queue const&) = delete;
+    intrusive_queue& operator=(intrusive_queue&&) = delete;
+
+    bool
+    empty() const noexcept
+    {
+        return head_ == nullptr;
+    }
+
+    void
+    push(T* w) noexcept
+    {
+        w->next_ = nullptr;
+        if(tail_)
+            tail_->next_ = w;
+        else
+            head_ = w;
+        tail_ = w;
+    }
+
+    void
+    splice(intrusive_queue& other) noexcept
+    {
+        if(other.empty())
+            return;
+        if(tail_)
+            tail_->next_ = other.head_;
+        else
+            head_ = other.head_;
+        tail_ = other.tail_;
+        other.head_ = nullptr;
+        other.tail_ = nullptr;
+    }
+
+    T*
+    pop() noexcept
+    {
+        if(!head_)
+            return nullptr;
+        T* w = head_;
+        head_ = head_->next_;
+        if(!head_)
+            tail_ = nullptr;
+        return w;
+    }
+};
+
+} // detail
 } // capy
 } // boost
